Works with --to-dir option of lein-new.
When --to-dir option of lein-new is used, the hypothesis that *name*
matches the last segment of --to-dir path does not hold.
Neither holds the hypothesis that the current directory is the parent
directory of the newly created project.

This is solved by using leiningen.new.templates/*dir* dynamic var
which is set by lein-newnew to the real project folder.